% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cer.R
\name{cer}
\alias{cer}
\title{Annualized Certainty Equivalent Return}
\usage{
cer(return, RRA, h)
}
\arguments{
\item{return}{something}

\item{RRA}{something}

\item{h}{something}
}
\description{
Annualized Certainty Equivalent Return
}
