% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_oos_gw.R
\name{is_oos_gw}
\alias{is_oos_gw}
\title{In-sample and out-of-sample Goyal and Welch (2008)}
\usage{
is_oos_gw(df, firstyear, cumsum = c("1", "0"))
}
\arguments{
\item{df}{tibble or dataframe with date, dependent variable and independent variable (variable)}

\item{firstyear}{Size of the windows}

\item{cumsum}{Cumulative sum of  sum squared error (sse)}
}
\description{
In-sample and out-of-sample Goyal and Welch (2008)
}
