% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivx_wald.R
\name{ivx_wald}
\alias{ivx_wald}
\title{IVX-Wald: Kostakis, A, T Magdalinos, and MP Stamatogiannis (2015)}
\usage{
ivx_wald(y, X, K, M_n, beta)
}
\arguments{
\item{y}{T-vector of return observations}

\item{X}{T-by-r data matrix of predictor observations}

\item{K}{Forecast horizon}

\item{M_n}{Bandwith parameter}

\item{beta}{Scalar in (0,1); specify 'close' to one for best performance}
}
\description{
IVX-Wald: Kostakis, A, T Magdalinos, and MP Stamatogiannis (2015)
}
