% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oos_portfolio.R
\name{oos_portfolio}
\alias{oos_portfolio}
\title{Title}
\usage{
oos_portfolio(start, end, ER_h, R_f_h, X, w_UB, w_LB, RRA, h, k)
}
\arguments{
\item{start}{something}

\item{end}{something}

\item{ER_h}{something}

\item{R_f_h}{something}

\item{X}{something}

\item{w_UB}{something}

\item{w_LB}{something}

\item{RRA}{something}

\item{h}{something}

\item{k}{something}
}
\description{
Title
}
