% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out_of_sample.R
\name{out_of_sample}
\alias{out_of_sample}
\title{Out-of-Sample}
\usage{
out_of_sample(y, X, y_2, h, start, end, space, k)
}
\arguments{
\item{y}{something}

\item{X}{something}

\item{y_2}{something}

\item{h}{something}

\item{start}{something}

\item{end}{something}

\item{space}{something}

\item{k}{something}
}
\description{
Out-of-Sample
}
