% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictors.R
\docType{data}
\name{predictors_annual}
\alias{predictors_annual}
\title{Goyal & Welch (2008) annual dataset}
\format{A data frame with 150 observations and 6 variables:}
\usage{
predictors_annual
}
\description{
Goyal & Welch (2008) annual dataset
}
\examples{
predictors_annual
}
\keyword{datasets}
