% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictors.R
\docType{data}
\name{predictors_quarterly}
\alias{predictors_quarterly}
\title{Nutrient definitions.}
\format{A data frame with 150 observations and 6 variables:}
\usage{
predictors_quarterly
}
\description{
Nutrient definitions.
}
\examples{
predictors_quarterly
}
\keyword{datasets}
