% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qLL_hat.R
\name{qLL_hat}
\alias{qLL_hat}
\title{TqLL_hat: Elliott and Muller (2006)}
\usage{
qLL_hat(y, X, Z, L)
}
\arguments{
\item{y}{T-vector of dependent variable observations}

\item{X}{T-by-k data matrix linked with potentially changing coefficients}

\item{Z}{T-by-d data matrix linked with constant coefficients (empty matrix if all coefficients allowed to change)}

\item{L}{lag truncation for Newey-West estimator}
}
\description{
TqLL_hat: Elliott and Muller (2006)
}
