% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_volatility.R
\name{return_volatility}
\alias{return_volatility}
\title{Return volatility}
\usage{
return_volatility(x1, x2)
}
\arguments{
\item{x1}{something}

\item{x2}{something}
}
\description{
Return volatility
}
