% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictors.R
\docType{data}
\name{predictors_monthly}
\alias{predictors_monthly}
\title{Food descriptions.}
\format{A data frame with 8463 observations and 14 variables:}
\usage{
predictors_monthly
}
\description{
Food descriptions.
}
\examples{
predictors_monthly
}
\keyword{datasets}
