% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictors.R
\docType{data}
\name{predictors_quarterly}
\alias{predictors_quarterly}
\title{Goyal & Welch (2008) quarterly dataset}
\format{A data frame with 150 observations and 6 variables:}
\usage{
predictors_quarterly
}
\description{
Goyal & Welch (2008) quarterly dataset
}
\examples{
predictors_quarterly
}
\keyword{datasets}
